backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
python-multipart==0.0.6
lxml==5.1.0
